# lib/openapi/spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Timesheet API",
    "version": "1.0.0",
    "description": "Complete REST API for the Timesheet workforce management system. Handles employee time tracking, scheduling, roster management, payroll and analytics.\n\n## Authentication\nUses cookie-based JWT authentication. Three separate auth layers:\n- Admin/user: login via /api/auth/login\n- Employee: login via /api/employee/login\n- Device: register via /api/device/register\n\n## Roles\n- super_admin: full system access\n- admin: manage employees and schedules\n- user: limited view-only dashboard access\n- employee: clock in/out and own timesheets\n\n## Base URL\nAll endpoints are prefixed with /api\n\n## Error Responses\nAll errors follow the format: { error: string, message?: string }\n\n## Cookie Auth\nAll protected endpoints require valid JWT cookie sent automatically via credentials: include on all requests."
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "message": {
            "type": "string",
            "example": "Invalid input data"
          },
          "details": {
            "type": "object"
          }
        },
        "required": ["error"]
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        },
        "required": ["success"]
      },
      "ObjectId": {
        "type": "string",
        "pattern": "^[0-9a-fA-F]{24}$",
        "example": "507f1f77bcf86cd799439011"
      },
      "DateTime": {
        "type": "string",
        "format": "date-time",
        "example": "2024-03-07T10:30:00.000Z"
      },
      "Date": {
        "type": "string",
        "format": "date",
        "example": "2024-03-07"
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "name": {
            "type": "string",
            "example": "Sales Associate"
          },
          "type": {
            "type": "string",
            "example": "role"
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "example": "#FF5733"
          },
          "description": {
            "type": "string",
            "example": "Responsible for customer service and sales"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/DateTime"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/rosters": {
      "get": {
        "tags": ["Rosters"],
        "summary": "Get all rosters with filtering options",
        "description": "Retrieves all roster schedules with optional filtering by date range and location. Returns roster summaries including week identifiers, publication status, total shifts, and total hours. Used by scheduling managers to view and manage workforce schedules across multiple locations and time periods. Supports pagination for large datasets.",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Date"
            }
          },
          {
            "name": "locationId",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of rosters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "$ref": "#/components/schemas/ObjectId"
                      },
                      "weekId": {
                        "type": "string",
                        "example": "2024-W10"
                      },
                      "startDate": {
                        "$ref": "#/components/schemas/Date"
                      },
                      "endDate": {
                        "$ref": "#/components/schemas/Date"
                      },
                      "isPublished": {
                        "type": "boolean",
                        "example": false
                      },
                      "totalShifts": {
                        "type": "number",
                        "example": 45
                      },
                      "totalHours": {
                        "type": "number",
                        "example": 360
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/locations/{locationId}/roles": {
      "get": {
        "tags": ["Locations"],
        "summary": "Get all roles available at specific location",
        "description": "Retrieves all job roles that are available and active at a specific location. Used for employee assignment, scheduling, and role management. Helps managers understand which positions are available at each location and supports workforce planning. Essential for location-based role assignments and scheduling.",
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Location roles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Location not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Locations"],
        "summary": "Add new role to location's available positions",
        "description": "Associates a job role with a specific location, making it available for employee assignments and scheduling at that location. Validates that the role exists and prevents duplicate assignments. Used for expanding available positions at locations and supporting business growth. Essential for location-based workforce management.",
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roleId": {
                    "$ref": "#/components/schemas/ObjectId"
                  },
                  "isActive": {
                    "type": "boolean",
                    "example": true
                  }
                },
                "required": ["roleId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role added to location successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Location not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}